+++ src/varinfo/varinfo.jl
@@ -1,0 +1,10 @@
+# Constants for caching
+const CACHERESET  = 0b00
+const CACHEIDCS   = 0b10
+const CACHERANGES = 0b01
+
+include("types.jl")
+include("utils.jl")
+include("indexing.jl")
+include("linking.jl")
+include("ad.jl")
