+++ test/Turing/contrib/inference/dynamichmc.jl
@@ -64,16 +64,16 @@ function AbstractMCMC.sample_init!(
     initialize_parameters!(spl; kwargs...)
 
     model(spl.state.vi, SampleFromUniform())
-    link!(spl.state.vi, spl)
+    link!(spl.state.vi, spl, model)
     l, dl = _lp(spl.state.vi[spl])
     while !isfinite(l) || !isfinite(dl)
         model(spl.state.vi, SampleFromUniform())
-        link!(spl.state.vi, spl)
+        link!(spl.state.vi, spl, model)
         l, dl = _lp(spl.state.vi[spl])
     end
 
     if spl.selector.tag == :default && !islinked(spl.state.vi, spl)
-        link!(spl.state.vi, spl)
+        link!(spl.state.vi, spl, model)
         model(spl.state.vi, spl)
     end
