+++ test/runtests.jl
@@ -8,8 +8,8 @@ using Random
 using Test
 
 dir = splitdir(splitdir(pathof(DynamicPPL))[1])[1]
-include(dir*"/test/Turing/Turing.jl")
-using .Turing
+#include(dir*"/test/Turing/Turing.jl")
+using Turing
 
 turnprogress(false)
